implementing new Code selection feature
- automatically compute a derived rating from
  a set of codes
- move modality and type information out of Exam into
  new ScanType table
- link CodeType to ScanType
- add image and user specific Codes which mirrors Ratings
- Ratings can be set manually independent of any Codes or
  set to the same value as a DerivedRating which is computed
  from a set of Codes
- Codes can be grouped so that a single CodeGroup penalty Value is
  applied to a Rating/DerivedRating when multiple Codes of the
  same CodeGroup are selected
- revise UI to include a QTableWidget populated by QCheckBox
  selectable Codes

+++ sql/upgrade/1.4/Exam.sql
@@ -1,6 +1,30 @@
-ALTER TABLE Exam
-DROP INDEX uqInterviewIdModalityIdTypeLaterality;
+DROP PROCEDURE IF EXISTS patch_Exam;
+DELIMITER //
+CREATE PROCEDURE patch_Exam()
+  BEGIN
 
-ALTER TABLE Exam
-ADD UNIQUE INDEX
-uqInterviewIdModalityIdTypeLaterality (InterviewId,ModalityId,Type,Laterality);
+    SELECT "Updating Exam table with new unique index" AS "";
+
+    SET @test = (
+      SELECT COUNT(*)
+      FROM information_schema.TABLE_CONSTRAINTS
+      WHERE TABLE_SCHEMA = DATABASE()
+      AND TABLE_NAME = "Exam"
+      AND CONSTRAINT_NAME = "uqInterviewIdModalityIdTypeLaterality" );
+
+    IF @test = 1 THEN
+
+      ALTER TABLE Exam
+      DROP INDEX uqInterviewIdModalityIdTypeLaterality;
+
+      ALTER TABLE Exam
+      ADD UNIQUE INDEX
+      uqInterviewIdModalityIdTypeLaterality (InterviewId,ModalityId,Type,Laterality);
+
+    END IF;
+
+  END //
+DELIMITER ;
+
+CALL patch_Exam();
+DROP PROCEDURE IF EXISTS patch_Exam;
